auth routing and reducer selectors refactor
