+++ src/services/redux/reducers/authReducer/authSelectors.js
@@ -1,0 +1,3 @@
+export const authStatusSelector = (state) => state.auth.status;
+
+export const authErrorSelector = (state) => state.auth.error;
