+++ src/services/redux/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import authReducer from './reducers/authReducer';
+import authReducer from './reducers/authReducer/authReducer';
 
 const store = configureStore({
   reducer: {
